pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "earthpy>=0.9.4",
    "geopandas>=1.0.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "rasterio>=1.4.3",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.14.0",
]
